pyproject.toml
[project]
name = "lpp-eval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jinja2>=3.1.4",
    "lpp-collector",
    "python-dotenv>=1.0.1",
    "python-redmine>=2.5.0",
    "flask>=3.0.0",
    "flask-sqlalchemy>=3.1.0",
    "apscheduler>=3.10.0",
]

[tool.uv.sources]
lpp-collector = { git = "https://github.com/f0reachARR/lpp_test.git" }
